requirements.txt
pandas
sodapy
pyspark
google-cloud-storage
google-cloud-bigquery
python-dotenv